refactor: move around some code for bundle shape management
